feat: remove the separate user info data loader and put in the loaders package
