+++ pkg/graph/resolvers/system_intake_notes.go
@@ -6,10 +6,10 @@ import (
 	"github.com/guregu/null"
 
 	"github.com/cmsgov/easi-app/pkg/appcontext"
-	"github.com/cmsgov/easi-app/pkg/dataloaders"
 	"github.com/cmsgov/easi-app/pkg/graph/model"
 	"github.com/cmsgov/easi-app/pkg/models"
 	"github.com/cmsgov/easi-app/pkg/storage"
+	"github.com/cmsgov/easi-app/pkg/storage/loaders"
 )
 
 // CreateSystemIntakeNote creates a system intake note.
@@ -71,7 +71,7 @@ func SystemIntakeNoteAuthor(obj *models.SystemIntakeNote) (*model.SystemIntakeNo
 func SystemIntakeNoteEditor(ctx context.Context, obj *models.SystemIntakeNote) (*models.UserInfo, error) {
 	var systemIntakeNoteEditorInfo *models.UserInfo
 	if obj.ModifiedBy != nil {
-		info, err := dataloaders.GetUserInfo(ctx, *obj.ModifiedBy)
+		info, err := loaders.GetUserInfo(ctx, *obj.ModifiedBy)
 		if err != nil {
 			return nil, err
 		}
